+++ src/main/java/com/finalpre/quickshare/dto/AdminSmtpTestRequest.java
@@ -1,0 +1,8 @@
+package com.finalpre.quickshare.dto;
+
+import lombok.Data;
+
+@Data
+public class AdminSmtpTestRequest {
+    private String toEmail;
+}